Update the peer.c to perform what the README says it can do
The original commit had a peer.c not performing features described in the original README. This was a mistake on my behalf caused by awful file management.
The Current file includes a great number of changes from the previous iteration. This includes 2 data structures for storing datas needed across threads for managing connections.

Signed-off-by: Diego G Esquivel <mpi.diegoe@gmail.com>